Add files via upload
This is an updated version for creating the co-occurrence networks that has both clearer code and some extra features, like including other words that can refer to persons that are not personal pronouns. Created to help research into the Psalms by PhD student Christiaan Erwich. More extensive description in the readme in the home of this GitHub